+++ src/main/kotlin/org/bashpile/core/BashpileState.kt
@@ -1,10 +1,10 @@
-package org.bashpile.core.bast
+package org.bashpile.core
 
 import org.bashpile.core.bast.types.TypeEnum
 import org.bashpile.core.bast.types.VariableTypeInfo
 
 /**
- * Sometimes the type of a node is known at creation, so the type may be in [BastNode] as well.
+ * Sometimes the type of a node is known at creation, so the type may be in [org.bashpile.core.bast.BastNode] as well.
  */
 class BashpileState {
     val stackframe: MutableList<VariableTypeInfo> = mutableListOf()
