+++ src/main/kotlin/org/bashpile/core/antlr/package-info.md
@@ -1,1 +1,2 @@
-For interfacing with Antlr4.
+For interfacing with Antlr4.  
+The AstConvertingVisitor listens for Antlr node visits and creates the initial BAST tree.
