+++ src/main/kotlin/org/bashpile/core/bast/BastNode.kt
@@ -20,7 +20,7 @@ typealias UnnestTuple = Pair<List<BastNode>, BastNode>
  * The base class of the BAST class hierarchy.
  * Converts this AST and children to the Bashpile text output via [render].
  * The root is created by the [AstConvertingVisitor].
- * Sometimes the type of a node isn't known at creation time, so the type is on the call stack at [BashpileState].
+ * Sometimes the type of a node isn't known at creation time, so the type is on the call stack at [org.bashpile.core.BashpileState].
  */
 abstract class BastNode(
     val children: List<BastNode>,
